Made path for makeapikeyfile relative, needs to be run from right dir now

+++ bin/makeapikeyfile_main.py
@@ -18,13 +18,7 @@ if __name__ == '__main__':
     
     args = parser.parse_args()
     
-    root = os.getenv( "MROOT")
-    
-    if root is None:
-        print( "You need to define MROOT to point to the github project root (MUGAlyser dir)")
-        sys.exit( 2 )
-    else:
-        root = os.path.join( root, "mugalyser")
+    root = os.path.join( "..", "mugalyser")
         
     apikey = args.apikey
     
@@ -33,17 +27,21 @@ if __name__ == '__main__':
         print( "Goto https://secure.meetup.com/meetup_api/key/ to get an API key for meetup.com")
         
     input_path = os.path.join( root, "apikey_template.py" )
-    output_path = os.path.join( root, "apikey.py" )
-    with open( input_path, "r") as infile:
-        with open( output_path, "w" ) as outfile:
-            outfile.write( "# Created: %s\n" % datetime.now())
-            for line in infile :
-                if line.startswith( "MEETUP_API_KEY"):
-                    outfile.write( "MEETUP_API_KEY=\"%s\"\n" % apikey )
-                    print( "%s -> %s" % ( apikey, output_path ))
-                else:
-                    outfile.write( line )
-        
+    
+    if os.path.isfile( input_path ) :
+        output_path = os.path.join( root, "apikey.py" )
+        with open( input_path, "r") as infile:
+            with open( output_path, "w" ) as outfile:
+                outfile.write( "# Created: %s\n" % datetime.now())
+                for line in infile :
+                    if line.startswith( "MEETUP_API_KEY"):
+                        outfile.write( "MEETUP_API_KEY=\"%s\"\n" % apikey )
+                        print( "%s -> %s" % ( apikey, output_path ))
+                    else:
+                        outfile.write( line )
+    else:
+        print( "No such file: '%s'" % input_path )
+        sys.exit( 2 )
     if apikey == "" :
         sys.exit( 2 )
     else:
